+++ Sources/Sentry/SentryANRTracker.m
@@ -3,7 +3,6 @@
 #import "SentryDispatchQueueWrapper.h"
 #import "SentryLog.h"
 #import "SentryThreadWrapper.h"
-#import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN