+++ Sources/Sentry/SentryClient.m
@@ -13,7 +13,6 @@
 #import "SentryEvent.h"
 #import "SentryException.h"
 #import "SentryFileManager.h"
-#import "SentryFrameRemover.h"
 #import "SentryGlobalEventProcessor.h"
 #import "SentryId.h"
 #import "SentryInAppLogic.h"
@@ -25,13 +24,9 @@
 #import "SentryMeta.h"
 #import "SentryNSError.h"
 #import "SentryOptions+Private.h"
-#import "SentryOptions.h"
 #import "SentryOutOfMemoryTracker.h"
-#import "SentryRandom.h"
 #import "SentrySDK+Private.h"
 #import "SentryScope+Private.h"
-#import "SentryScope.h"
-#import "SentrySpan.h"
 #import "SentryStacktraceBuilder.h"
 #import "SentryThreadInspector.h"
 #import "SentryTraceState.h"
@@ -42,11 +37,12 @@
 #import "SentryUser.h"
 #import "SentryUserFeedback.h"
 
+
 #if SENTRY_HAS_UIKIT
 #    import <UIKit/UIKit.h>
-
+#    import "SentryScreenshot.h"
 #endif
-#import "SentryScreenshot.h"
+
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -391,6 +387,11 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
     [self.fileManager storeEnvelope:envelope];
 }
 
+- (void)recordLostEvent:(SentryDataCategory)category reason:(SentryDiscardReason)reason
+{
+    [self.transport recordLostEvent:category reason:reason];
+}
+
 - (SentryEvent *_Nullable)prepareEvent:(SentryEvent *)event
                              withScope:(SentryScope *)scope
                 alwaysAttachStacktrace:(BOOL)alwaysAttachStacktrace
@@ -419,6 +420,7 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
     if (eventIsNotATransaction && [self isSampled:self.options.sampleRate]) {
         [SentryLog logWithMessage:@"Event got sampled, will not send the event"
                          andLevel:kSentryLevelDebug];
+        [self recordLostEvent:kSentryDataCategoryError reason:kSentryDiscardReasonSampleRate];
         return nil;
     }
 
@@ -492,16 +494,23 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
     }
 
     event = [self callEventProcessors:event];
+    if (event == nil) {
+        [self recordLost:eventIsNotATransaction reason:kSentryDiscardReasonEventProcessor];
+    }
 
-    if (nil != self.options.beforeSend) {
+    if (event != nil && nil != self.options.beforeSend) {
         event = self.options.beforeSend(event);
+
+        if (event == nil) {
+            [self recordLost:eventIsNotATransaction reason:kSentryDiscardReasonBeforeSend];
+        }
     }
 
     if (isCrashEvent && nil != self.options.onCrashedLastRun && !SentrySDK.crashedLastRunCalled) {
         // We only want to call the callback once. It can occur that multiple crash events are
         // about to be sent.
-        self.options.onCrashedLastRun(event);
         SentrySDK.crashedLastRunCalled = YES;
+        self.options.onCrashedLastRun(event);
     }
 
     return event;
@@ -627,6 +636,16 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
     event.context = context;
 }
 
+- (void)recordLost:(BOOL)eventIsNotATransaction reason:(SentryDiscardReason)reason
+{
+    if (eventIsNotATransaction) {
+        [self recordLostEvent:kSentryDataCategoryError reason:reason];
+    } else {
+        [self recordLostEvent:kSentryDataCategoryTransaction reason:reason];
+    }
+}
+
+
 - (void)attachScreenshots:(SentryScope *)scope
 {
 #if SENTRY_HAS_UIKIT
@@ -649,6 +668,7 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
     }
 #endif
 }
+
 
 @end