+++ Sources/Sentry/SentryDefaultCurrentDateProvider.m
@@ -1,5 +1,4 @@
 #import "SentryDefaultCurrentDateProvider.h"
-#import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN