+++ Sources/Sentry/SentryFileManager.m
@@ -1,18 +1,15 @@
 #import "SentryFileManager.h"
 #import "NSDate+SentryExtras.h"
 #import "SentryAppState.h"
-#import "SentryDefaultCurrentDateProvider.h"
+#import "SentryDataCategoryMapper.h"
 #import "SentryDsn.h"
 #import "SentryEnvelope.h"
-#import "SentryEnvelopeItemType.h"
-#import "SentryError.h"
 #import "SentryEvent.h"
 #import "SentryFileContents.h"
 #import "SentryLog.h"
 #import "SentryMigrateSessionInit.h"
 #import "SentryOptions.h"
 #import "SentrySerialization.h"
-#import "SentrySession+Private.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -29,6 +26,7 @@ SentryFileManager ()
 @property (nonatomic, copy) NSString *appStateFilePath;
 @property (nonatomic, assign) NSUInteger currentFileCounter;
 @property (nonatomic, assign) NSUInteger maxEnvelopes;
+@property (nonatomic, weak) id<SentryFileManagerDelegate> delegate;
 
 @end
 
@@ -77,6 +75,11 @@ SentryFileManager ()
         self.maxEnvelopes = options.maxCacheItems;
     }
     return self;
+}
+
+- (void)setDelegate:(id<SentryFileManagerDelegate>)delegate
+{
+    _delegate = delegate;
 }
 
 - (void)deleteAllFolders
@@ -212,9 +215,27 @@ SentryFileManager ()
             [[NSMutableArray alloc] initWithArray:[envelopeFilePaths copy]];
         [envelopePathsCopy removeObjectAtIndex:i];
 
-        [SentryMigrateSessionInit migrateSessionInit:envelopeFilePath
-                                    envelopesDirPath:self.envelopesPath
-                                   envelopeFilePaths:envelopePathsCopy];
+        NSData *envelopeData = [[NSFileManager defaultManager] contentsAtPath:envelopeFilePath];
+        SentryEnvelope *envelope = [SentrySerialization envelopeWithData:envelopeData];
+
+        BOOL didMigrateSessionInit =
+            [SentryMigrateSessionInit migrateSessionInit:envelope
+                                        envelopesDirPath:self.envelopesPath
+                                       envelopeFilePaths:envelopePathsCopy];
+
+        for (SentryEnvelopeItem *item in envelope.items) {
+            SentryDataCategory rateLimitCategory =
+                [SentryDataCategoryMapper mapEnvelopeItemTypeToCategory:item.header.type];
+
+            // When migrating the session init, the envelope to delete still contains the session
+            // migrated to another envelope. Therefore, the envelope item is not deleted but
+            // migrated.
+            if (didMigrateSessionInit && rateLimitCategory == kSentryDataCategorySession) {
+                continue;
+            }
+
+            [_delegate envelopeItemDeleted:rateLimitCategory];
+        }
 
         [self removeFileAtPath:envelopeFilePath];
     }