+++ Tests/SentryTests/ClearTestState.swift
@@ -19,4 +19,5 @@ func clearTestState() {
     #endif
     
     SentryDependencyContainer.reset()
+    Dynamic(SentryGlobalEventProcessor.shared()).removeAllProcessors()
 }